Move ExecResult to Synapse
