+++ src/synapse.py
@@ -13,7 +13,6 @@ import yaml
 from ops.pebble import Check, ExecError, PathError
 
 from charm_state import CharmState
-from charm_types import ExecResult
 from constants import (
     CHECK_READY_NAME,
     COMMAND_MIGRATE_CONFIG,
@@ -25,6 +24,20 @@ from constants import (
 from exceptions import CommandMigrateConfigError, ServerNameModifiedError
 
 logger = logging.getLogger(__name__)
+
+
+class ExecResult(typing.NamedTuple):
+    """A named tuple representing the result of executing a command.
+
+    Attributes:
+        exit_code: The exit status of the command (0 for success, non-zero for failure).
+        stdout: The standard output of the command as a string.
+        stderr: The standard error output of the command as a string.
+    """
+
+    exit_code: int
+    stdout: str
+    stderr: str
 
 
 class Synapse:
